+++ imagedata/download.go
@@ -11,6 +11,7 @@ import (
 
 	"github.com/imgproxy/imgproxy/v3/config"
 	"github.com/imgproxy/imgproxy/v3/ierrors"
+	"github.com/imgproxy/imgproxy/v3/security"
 
 	azureTransport "github.com/imgproxy/imgproxy/v3/transport/azure"
 	fsTransport "github.com/imgproxy/imgproxy/v3/transport/fs"
@@ -38,6 +39,11 @@ var (
 )
 
 const msgSourceImageIsUnreachable = "Source image is unreachable"
+
+type DownloadOptions struct {
+	Header    http.Header
+	CookieJar *cookiejar.Jar
+}
 
 type ErrorNotModified struct {
 	Message string
@@ -142,7 +148,7 @@ func BuildImageRequest(imageURL string, header http.Header, jar *cookiejar.Jar) 
 	if _, ok := enabledSchemes[req.URL.Scheme]; !ok {
 		return nil, ierrors.New(
 			404,
-			fmt.Sprintf("Unknown sheme: %s", req.URL.Scheme),
+			fmt.Sprintf("Unknown scheme: %s", req.URL.Scheme),
 			msgSourceImageIsUnreachable,
 		)
 	}
@@ -173,8 +179,8 @@ func SendRequest(req *http.Request) (*http.Response, error) {
 	return res, nil
 }
 
-func requestImage(imageURL string, header http.Header, jar *cookiejar.Jar) (*http.Response, error) {
-	req, err := BuildImageRequest(imageURL, header, jar)
+func requestImage(imageURL string, opts DownloadOptions) (*http.Response, error) {
+	req, err := BuildImageRequest(imageURL, opts.Header, opts.CookieJar)
 	if err != nil {
 		return nil, err
 	}
@@ -205,13 +211,13 @@ func requestImage(imageURL string, header http.Header, jar *cookiejar.Jar) (*htt
 	return res, nil
 }
 
-func download(imageURL string, header http.Header, jar *cookiejar.Jar) (*ImageData, error) {
+func download(imageURL string, opts DownloadOptions, secopts security.Options) (*ImageData, error) {
 	// We use this for testing
 	if len(redirectAllRequestsTo) > 0 {
 		imageURL = redirectAllRequestsTo
 	}
 
-	res, err := requestImage(imageURL, header, jar)
+	res, err := requestImage(imageURL, opts)
 	if res != nil {
 		defer res.Body.Close()
 	}
@@ -234,7 +240,7 @@ func download(imageURL string, header http.Header, jar *cookiejar.Jar) (*ImageDa
 		contentLength = 0
 	}
 
-	imgdata, err := readAndCheckImage(body, contentLength)
+	imgdata, err := readAndCheckImage(body, contentLength, secopts)
 	if err != nil {
 		return nil, ierrors.Wrap(err, 0)
 	}